// src/app/flows/cases/cases-create-casefile/cases-create-casefile-order-details/interfaces/cases-create-casefile-order-details-form-data.interface.ts
export interface ICasesCreateCasefileOrderDetailsFormData {
  create_casefile_order_details_application_id: number | string | null;
  create_casefile_order_details_court: string | null;
  create_casefile_order_details_date_order_made: string | null;
  create_casefile_order_details_payment_frequency: string | null;
  create_casefile_order_details_date_arrears_last_updated: string | null;
}